Build each cross-target fixture once instead of six times (#952)
* Build each cross-target fixture once and share the legs across all three corpus gates

* Repoint the last interp CLAUDE.md reference at the relocated ledger gate
