+++ crates/almide-interp/CLAUDE.md
@@ -6,9 +6,9 @@ A tree-walking interpreter over the **pre-codegen** `IrProgram` — the third le
 
 The cross-target gate (`spec/wasm_cross/*.almd`, enforced by `tests/wasm_runtime_test.rs::wasm_cross_target_spec`) compares the **native** and **WASM** backends and requires byte-identical `(stdout, stderr, exit)`. That 2-way vote is structurally blind to a *both-wrong-the-same-way* bug: if codegen and the WASM emitter share a lowering pass that is wrong, both agree and the gate stays green while the behaviour is wrong.
 
-This crate adds a third, independent judge. It evaluates the IR at the cut point **after** `lower → optimize → mono → ir_link` but **before** any target lowering — almide-codegen's Rust passes (`Clone` / `StdlibLowering` / `BoxDeref` / …) and almide-mir's wasm lowering alike (the codegen wasm passes named here previously — ClosureConversion, Perceus — were dead code retired in #930). So it shares *none* of either backend's target passes. The ~22 codegen-inserted `IrExprKind` variants are unreachable here and `eval.rs` asserts them `unreachable!` to guard the boundary.
+This crate adds a third, independent judge. It evaluates the IR at the cut point **after** `lower → optimize → mono → ir_link` but **before** any of `almide-codegen`'s target-lowering passes (`ClosureConversion` / `Perceus` / `StdlibLowering` / `IterChain` / …). So it shares *none* of codegen's target passes with either backend. The ~22 codegen-inserted `IrExprKind` variants are unreachable here and `eval.rs` asserts them `unreachable!` to guard the boundary.
 
-`tests/interp_cross_target_test.rs::interp_cross_target_spec` is the 3-way harness:
+`tests/wasm_runtime_test.rs::interp_cross_target_spec` (root binary — it shares one build of each fixture with the equivalence and wasm-opt gates) is the 3-way harness:
 - `interp == native == wasm` → corroborated by a spec that cannot share a codegen bug.
 - `native != wasm` → a backend split **owned by the 2-way `@xt-allow` gate**; this harness only logs which backend the interp sides with (tie-break diagnostic), it does NOT fail.
 - `native == wasm` but interp dissents → **LOUD failure** (`BOTH-BACKENDS-WRONG` banner). Diagnose: fix the interpreter, or you just found a both-backends bug the 2-way gate is blind to.
@@ -24,18 +24,15 @@ This crate adds a third, independent judge. It evaluates the IR at the cut point
 
 ## Coverage model — does a NEW stdlib fn get covered automatically?
 
-**A pure-scalar SELF-HOSTED fn: yes, automatically. Everything else needs manual glue, or it becomes an honest skip.** Dispatch is three hand-maintained surfaces plus one automatic tier, tried in this order per `(module, func)`:
+**No. A new stdlib fn needs manual glue here, or it becomes an honest skip.** Dispatch is three hand-maintained surfaces keyed by `(module, func)`:
 
 1. `dispatch.rs::is_hof` — the closure-taking combinators (the allowlist).
 2. `hofs.rs::eval_container_op` — non-HOF structural container ops.
-3. `bridge.rs::dispatch` — scalar `int`/`float`/`string`/`math`/`bool` fns, plus the **scalar prim floor** (`prim.band`, the f64/f32 families, …) the tier below consumes.
-4. `stdlib_pool.rs` — **the automatic tier**: the self-hosted stdlib bodies from `almide_lang::self_host_registry` (the SAME table + sources the wasm renderer links), lowered once per process and evaluated as ordinary fns. Register a new self-hosted fn in the registry (which its wasm build already requires) and the oracle covers it with no glue here — evaluability is decided per fn at call time by what the body touches: a body reaching a heap/effect prim (`prim.handle`, `prim.store8`, `prim.fd_write`…) abstains with that prim named. There is deliberately no hand-maintained "these modules are scalar" list.
+3. `bridge.rs::dispatch` — scalar `int`/`float`/`string`/`math`/`bool` fns.
 
-Anything not matched falls through to `Flow::Unsupported(...)`, which the 3-way gate records as a **data-driven, reasoned skip** (printed with its concrete reason; there is no hardcoded skip-list). Coverage shrinkage is no longer silent: the abstain set is audited against the committed `interp-abstain-ledger.txt` by `tests/interp_cross_target_test.rs::interp_abstain_ledger` (backend-free, never self-skips). A fixture abstaining without a ledger entry fails CI — so a new stdlib fn whose fixture the interp cannot run forces either interp glue or a reviewed ledger entry in the same PR — and a stale entry whose fixture now evaluates also fails (the ledger only shrinks; CG-1 gap-audit ratchet, issue #564).
+Anything not matched by one of these falls through to `Flow::Unsupported(...)`, which the 3-way gate records as a **data-driven, reasoned skip** (printed with its concrete reason; there is no hardcoded skip-list). Coverage shrinkage is no longer silent: the abstain set is audited against the committed `interp-abstain-ledger.txt` by `tests/wasm_runtime_test.rs::interp_abstain_ledger` (backend-free, never self-skips — it deliberately does NOT read the shared corpus table, since touching it would trigger 318 native+wasm builds and destroy exactly that property). A fixture abstaining without a ledger entry fails CI — so a new stdlib fn whose fixture the interp cannot run forces either interp glue or a reviewed ledger entry in the same PR — and a stale entry whose fixture now evaluates also fails (the ledger only shrinks; CG-1 gap-audit ratchet, issue #564). Widening this bridge surface (the partial json/regex/fan/map-constructor/bytes/string surface) is the highest-leverage follow-up: each added fn converts a skip into a real third vote and deletes ledger lines.
 
-When you add a stdlib fn and want the oracle to cover it: if it is self-hosted and scalar, the registry entry it already needs for wasm is enough. Only a true `@intrinsic` (body `_`) or a container op needs an arm on surfaces 1–3, mirroring the native runtime fn's behaviour exactly (the Rust-std behaviour IS the oracle for the scalar fns). Do NOT hand-mirror a self-hosted body in `bridge.rs` — that is a second derivation of one definition, and the two WILL drift (the `Int | Bool` capture-class subset and the E024 dual peel were both this disease); if the body cannot evaluate, widen the prim floor instead. Do NOT path-depend on `almide_rt` — that crate pulls in rustls/webpki (network/TLS) and effectful surface the interp does not need.
-
-The bridge arms that PREDATE the pool still win by dispatch order, so their vote provenance is unchanged; flipping an arm over to its pool body (and deleting the arm) is verifiable follow-up work, judged arm by arm by the 3-way gate.
+When you add a stdlib fn and want the oracle to cover it: add the arm to the right surface above, mirroring the native runtime fn's behaviour exactly (the Rust-std behaviour IS the oracle for the scalar fns). Do NOT path-depend on `almide_rt` for this — that crate pulls in rustls/webpki (network/TLS) and effectful surface the interp does not need.
 
 ## Rules
