+++ crates/almide-interp/Cargo.toml
@@ -16,7 +16,9 @@ almide-optimize = { path = "../almide-optimize" }
 # The test battery lowers .almd source to IR through the same public frontend
 # functions the driver uses, then interprets. No fixtures, all inline.
 #
-# `interp_cross_target_test.rs` additionally drives the native and WASM backends
-# as subprocesses (the `almide` binary built into the workspace target dir) to
-# run the 3-way oracle over `spec/wasm_cross/*.almd`, so it needs a scratch dir.
-tempfile = "3"
+# The corpus gates that used to live here — the 3-way oracle and the abstain
+# ledger over `spec/wasm_cross/*.almd` — moved to the root binary
+# (`tests/wasm_runtime_test.rs`), where they share ONE build of each fixture
+# with the equivalence and wasm-opt gates instead of compiling it again. That
+# also retired this crate's copies of the native/wasm capture helpers, and with
+# them the `tempfile` scratch dir they needed.
